structs: Prevent crash at app exit
Protect destructions of elements in GetAccelStructGeomHostAllocMap.
It should not be needed to check for emptyness of the map, but testing
showed doing so fixes a crash that can occur at application exit.

+++ include/vulkan/utility/vk_concurrent_unordered_map.hpp
@@ -175,7 +175,7 @@ class unordered_map {
 
     bool empty() const {
         bool result = 0;
-        for (int h = 0; h < BUCKETS; ++h) {
+        for (size_t h = 0; h < BUCKETS; ++h) {
             ReadLockGuard lock(locks[h].lock);
             result |= maps[h].empty();
         }
